v2.2

+++ EsDnevnik/Forma.cs
@@ -18,43 +18,54 @@ namespace EsDnevnik
             InitializeComponent();
         }
 
-        private void Skolska_Godina_Load(object sender, EventArgs e)
+        private void Osvezi(string tabela)
         {
             podaci = new DataTable();
             podaci = Konekcija.Unos("SELECT * FROM " + tabela);
 
             if (tabela == "Predmet")
             {
+                textBox3.Visible = true;
+                label3.Visible = true;
+                label2.Text = tabela;
+                dataGridView1.Columns["Smer"].Visible = false;
                 for (int i = 0; i < podaci.Rows.Count; i++)
                 {
                     dataGridView1.Rows.Add();
-                    dataGridView1.Rows[i].Cells["ID"].Value = Convert.ToString(podaci.Rows[i]["id"]);
+                    dataGridView1.Rows[i].Cells["IDD"].Value = Convert.ToString(podaci.Rows[i]["id"]);
                     dataGridView1.Rows[i].Cells["Naziv"].Value = Convert.ToString(podaci.Rows[i]["naziv"]);
                     dataGridView1.Rows[i].Cells["Razred"].Value = Convert.ToString(podaci.Rows[i]["razred"]);
                 }
-                
+
             }
             else if (tabela == "Smer")
             {
+                label2.Text = tabela;
                 dataGridView1.Columns["Naziv"].Visible = false;
                 dataGridView1.Columns["Razred"].Visible = false;
 
                 for (int i = 0; i < podaci.Rows.Count; i++)
                 {
                     dataGridView1.Rows.Add();
-                    dataGridView1.Rows[i].Cells["ID"].Value = Convert.ToString(podaci.Rows[i]["id"]);
+                    dataGridView1.Rows[i].Cells["IDD"].Value = Convert.ToString(podaci.Rows[i]["id"]);
                     dataGridView1.Rows[i].Cells["Smer"].Value = Convert.ToString(podaci.Rows[i]["naziv"]);
                 }
             }
             else
             {
+                label2.Text = "Skolska godina";
                 dataGridView1.Columns["Naziv"].Visible = false;
                 dataGridView1.Columns["Razred"].Visible = false;
                 dataGridView1.Columns["Smer"].Visible = false;
-                dataGridView1.Columns["ID"].Visible = false;
+                dataGridView1.Columns["IDD"].Visible = false;
                 dataGridView1.DataSource = podaci;
                 dataGridView1.Columns["id"].ReadOnly = true;
             }
+        }
+
+        private void Skolska_Godina_Load(object sender, EventArgs e)
+        {
+            Osvezi(tabela);
         }
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
@@ -65,24 +76,27 @@ namespace EsDnevnik
                 {
                     if (MessageBox.Show("Da li ste sigurni da zelite da obrisete ove podatake?", "EsDnevnik", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                     {
-                        int kurac = e.RowIndex;
                         int indeks;
-                        indeks = Convert.ToInt32(dataGridView1.Rows[e.RowIndex].Cells["id"].Value);
-                        menjanja = new SqlCommand();
-                        menjanja.CommandText = ("DELETE FROM " + tabela + " WHERE id = " + indeks);
+                        if (tabela == "Skolska_godina")
+                        {
+                            indeks = Convert.ToInt32(dataGridView1.Rows[e.RowIndex].Cells["id"].Value);
+                            menjanja = new SqlCommand();
+                            menjanja.CommandText = ("DELETE FROM " + tabela + " WHERE id = " + indeks);
 
-                        SqlConnection con = new SqlConnection(Konekcija.Veza());
-                        con.Open();
-                        menjanja.Connection = con;
-                        menjanja.ExecuteNonQuery();
-                        con.Close();
+                            SqlConnection con = new SqlConnection(Konekcija.Veza());
+                            con.Open();
+                            menjanja.Connection = con;
+                            menjanja.ExecuteNonQuery();
+                            con.Close();
 
-                        dataGridView1.Rows.RemoveAt(e.ColumnIndex);
+                            dataGridView1.Rows.RemoveAt(e.ColumnIndex);
 
-                        podaci = new DataTable();
-                        podaci = Konekcija.Unos("SELECT * FROM " + tabela);
-                        dataGridView1.DataSource = podaci;
-                        dataGridView1.Columns["id"].ReadOnly = true;
+                            Osvezi(tabela);
+                        }
+                        else
+                        {
+                            indeks = Convert.ToInt32(dataGridView1.Rows[e.RowIndex].Cells["ID"].Value);
+                        }
                     }
                 }
                 catch (Exception ex)
